+++ src/main/java/com/almostreliable/summoningrituals/recipe/condition/custom/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.recipe.condition.custom;
+
+import org.lwjgl.system.NonnullDefault;